+++ storage/timehash/timehash.c
@@ -158,7 +158,7 @@ TOKEN timehash_store(const ARTHANDLE article, const STORAGECLASS class) {
     }
     if (i == 0x10000) {
 	SMseterror(SMERR_UNDEFINED, NULL);
-	syslog(L_ERROR, "timehash: all sequence numbers for the time and class are reserved %lu %d", now, class);
+	syslog(L_ERROR, "timehash: all sequence numbers for the time and class are reserved %lu %d", (unsigned long)now, class);
 	token.type = TOKEN_EMPTY;
 	free(path);
 	return token;
